refactor: splits settings e2e tests
